Change calculate to accept float32

+++ cmd/calculate/entry.go
@@ -3,7 +3,7 @@ package calculate
 import "errors"
 
 // Calculate returns the calculated result requested
-func Calculate(a int, b int, operation string) (result float32, err error) {
+func Calculate(a float32, b float32, operation string) (result float32, err error) {
 
 	switch operation {
 	case "ADD":
@@ -20,18 +20,18 @@ func Calculate(a int, b int, operation string) (result float32, err error) {
 	return result, nil
 }
 
-func add(a int, b int) float32 {
-	return float32(a) + float32(b)
+func add(a float32, b float32) float32 {
+	return a + b
 }
 
-func subtract(a int, b int) float32 {
-	return float32(a) - float32(b)
+func subtract(a float32, b float32) float32 {
+	return a - b
 }
 
-func multiply(a int, b int) float32 {
-	return float32(a) * float32(b)
+func multiply(a float32, b float32) float32 {
+	return a * b
 }
 
-func divide(a int, b int) float32 {
-	return float32(a) / float32(b)
+func divide(a float32, b float32) float32 {
+	return a / b
 }
